macos fixed

+++ tools/workspace/vtk/image/vtk_common.py
@@ -33,6 +33,14 @@ class VtkArchive:
     sha256_sum_path: Path
 
 
+def system_is_linux():
+    return platform.system() == "Linux"
+
+
+def system_is_macos():
+    return platform.system() == "Darwin"
+
+
 def architecture() -> str:
     machine = platform.machine()
     assert machine in {
@@ -83,14 +91,6 @@ def _rlocation(relative_path: str) -> Path:
     )
 
 
-def system_is_linux():
-    return platform.system() == "Linux"
-
-
-def system_is_macos():
-    return platform.system() == "Darwin"
-
-
 def vtk_git_ref() -> str:
     return "d706250a1422ae1e7ece0fa09a510186769a5fec"
 
@@ -131,17 +131,35 @@ def vtk_archive_name(source_dir: Path) -> str:
 
 def build_vtk(package_tree: PackageTree, configure_args: list[str]):
     """Build VTK with the specified CMake configure arguments using Ninja."""
-    # Always start with a clean CMake build tree.
-    if package_tree.build_dir.is_dir():
-        shutil.rmtree(package_tree.build_dir)
-    elif package_tree.build_dir.is_file():
-        package_tree.build_dir.unlink()
-    package_tree.build_dir.mkdir(parents=True, exist_ok=False)
+    # Always start with a clean CMake build tree on Linux.  On macOS, reuse the
+    # existing build tree and re-run CMake configure.
+    if system_is_linux():
+        if package_tree.build_dir.is_dir():
+            shutil.rmtree(package_tree.build_dir)
+        elif package_tree.build_dir.is_file():
+            package_tree.build_dir.unlink()
+        exist_ok = False
+    elif system_is_macos():
+        exist_ok = True
+    package_tree.build_dir.mkdir(parents=True, exist_ok=exist_ok)
+
+    # Always start with a clean install tree.  Do this before building to save
+    # time when there are errors.
+    if package_tree.install_dir.is_dir():
+        shutil.rmtree(package_tree.install_dir)
+    elif package_tree.install_dir.is_file():
+        package_tree.install_dir.unlink()
+    if package_tree.install_dir.exists():
+        raise RuntimeError(
+            f"{package_tree.install_dir} exists but could not be removed, "
+            "please delete manually and rerun."
+        )
+
+    # (Re)run CMake configure.
     subprocess.check_call(
         [
             "cmake",
-            # TODO(svenevs): this causes issues on macOS.
-            # "-Werror",
+            "-Werror=dev",
             "-G",
             "Ninja",
             f"-DCMAKE_INSTALL_PREFIX:PATH={package_tree.install_dir}",
@@ -153,6 +171,7 @@ def build_vtk(package_tree: PackageTree, configure_args: list[str]):
             str(package_tree.source_dir),
         ]
     )
+    # Run CMake build / install.
     subprocess.check_call(
         [
             "cmake",
@@ -170,7 +189,7 @@ def package_vtk(package_tree: PackageTree) -> VtkArchive:
     tar_gz_path = package_tree.root / archive_name
     # Creating the .tgz can take a little while, display something to the user
     # so it's clear the script isn't stuck.
-    print(f"==> Compressing {package_tree.install_dir} to {tar_gz_path}.")
+    print(f"=> Compressing {package_tree.install_dir} to {tar_gz_path}.")
     with tarfile.open(tar_gz_path, "w:gz") as tgz:
         # NOTE: the second argument to `add` is what to rename things as, by
         # using the empty string it means that all components are stripped.
@@ -179,7 +198,7 @@ def package_vtk(package_tree: PackageTree) -> VtkArchive:
 
     # Create the accompanying sha256sum verification file.
     sha256_sum_path = tar_gz_path.parent / f"{tar_gz_path.name}.sha256"
-    print(f"==> Computing sha256sum of {tar_gz_path} into {sha256_sum_path}.")
+    print(f"=> Computing sha256sum of {tar_gz_path} into {sha256_sum_path}.")
     sha256 = hashlib.sha256()
     with open(tar_gz_path, "rb", buffering=0) as tgz:
         while True:
